Fixed scoring error and duplication error
The app was scoring1 point per restaurant star rating and 1 point per matched category, and this was throwing off the results. Matching categories should be the primary factor. Changing the per category score to 5 fixed this. But it made it far more common that duplicate restaurants are returned given franchises exist. Had to add a method to skip over duplicate decisions.